udelearn_app_back/requirements.txt
autopep8==2.3.2
blinker==1.9.0
click==8.2.1
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycodestyle==2.14.0
python-decouple==3.8
simplejson==3.20.1
Werkzeug==3.1.3
